requirements.txt
sentence-transformers>=3.0.0
langchain==0.1.20
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone>=0.1.0
langchain_community
langchain_openai
langchain_experimental
-e .